Modularize forms for IDSR steps
